+++ hexital/core/indicator.py
@@ -10,6 +10,7 @@ from hexital.core.candle import Candle
 from hexital.core.candle_manager import CandleManager
 from hexital.core.candlestick_type import CandlestickType
 from hexital.utils.candles import (
+    candles_average,
     candles_sum,
     get_readings_period,
     reading_by_candle,
@@ -321,6 +322,21 @@ class Indicator(ABC):
         include_latest: bool = True,
     ) -> float | None:
         return candles_sum(
+            self.candles,
+            name if name else self.name,
+            length,
+            index if index is not None else self._active_index,
+            include_latest,
+        )
+
+    def candles_average(
+        self,
+        length: int = 1,
+        name: Optional[str] = None,
+        index: Optional[int] = None,
+        include_latest: bool = True,
+    ) -> float | None:
+        return candles_average(
             self.candles,
             name if name else self.name,
             length,